move From<DeniedHttpResponse> for DirectResponse to the auth module
